Fix IPv6 support.
In Logstash 5.x and below, JRuby 1.7.x is used, and
UDPSocket.new(Socket::AF_INET) somehow accidentally works with IPv6.

Logstash 6.x uses JRuby 9.x where this now behaves correctly (AF_INET is
ipv4 only), but triggers a bug where ipv6 was accidentally disabled in
the udp input. Oops!

This change is small and adds tests to cover the
default scenario that 127.0.0.1 and ::1 both work for the default.